requirements.txt
fastapi
tensorflow
numpy
uvicorn
gunicorn
matplotlib
pillow